+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   File upload endpoint request and response body
+-   Change dependency from `multer` and `multer-s3` to `express-fileupload`
+
 ### Added
 
 -   Convenience method `yarn reset-db` for clean reset of db schema.
@@ -28,6 +33,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   DB queries for locations, auth etc
 
 ## [v0.63.0]
+
+### Changed
+
+-   Build file changed back to `dist/server.js`
 
 ## [v0.62.0]